Set up make file to support linting, app dependency installation, building the app and building the docker image.
