+++ .bingo/.gitignore
@@ -1,0 +1,13 @@
+
+# Ignore everything
+*
+
+# But not these files:
+!.gitignore
+!*.mod
+!*.sum
+!README.md
+!Variables.mk
+!variables.env
+
+*tmp.mod
