+++ .bingo/bingo.mod
@@ -1,0 +1,5 @@
+module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
+
+go 1.19
+
+require github.com/bwplotka/bingo v0.6.0
