+++ .bingo/golangci-lint.mod
@@ -1,0 +1,5 @@
+module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
+
+go 1.17
+
+require github.com/golangci/golangci-lint v1.45.2 // cmd/golangci-lint
