+++ .bingo/variables.env
@@ -1,0 +1,14 @@
+# Auto generated binary variables helper managed by https://github.com/bwplotka/bingo v0.6. DO NOT EDIT.
+# All tools are designed to be build inside $GOBIN.
+# Those variables will work only until 'bingo get' was invoked, or if tools were installed via Makefile's Variables.mk.
+GOBIN=${GOBIN:=$(go env GOBIN)}
+
+if [ -z "$GOBIN" ]; then
+	GOBIN="$(go env GOPATH)/bin"
+fi
+
+
+BINGO="${GOBIN}/bingo-v0.6.0"
+
+GOLANGCI_LINT="${GOBIN}/golangci-lint-v1.50.0"
+
