+++ go.mod
@@ -3,10 +3,10 @@ module github.com/nik-johnson-net/prometheus-envoy
 go 1.13
 
 require (
-	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
 	github.com/nik-johnson-net/go-envoy v1.0.0
 	github.com/nik-johnson-net/prometheus-proxy v1.0.0
-	github.com/prometheus/client_golang v1.6.0
-	github.com/prometheus/common v0.10.0 // indirect
-	golang.org/x/sys v0.0.0-20200519105757-fe76b779f299 // indirect
+	github.com/prometheus/client_golang v1.12.1
+	golang.org/x/sys v0.0.0-20220915200043-7b5979e65e41 // indirect
+	google.golang.org/protobuf v1.28.0 // indirect
 )
